Add files via upload
This data set contains records relevant to a direct marketing campaign of a Portuguese banking institution. The marketing campaign was executed through phone calls. Often, more than one call needs to be made to a single client before they either decline or agree to a term deposit subscription. The classification goal is to predict if the client will subscribe (yes/no) to the term deposit (variable y).

This is a modified version of the classic bank marketing data set originally shared in the UCI Machine Learning Repository. There are four datasets available on UCI's repository:
1) bank-additional-full.csv with all examples (41188) and 20 inputs, ordered by date (from May 2008 to November 2010), very close to the data analyzed in [Moro et al., 2014]
2) bank-additional.csv with 10% of the examples (4119), randomly selected from 1), and 20 inputs.
3) bank-full.csv with all examples and 17 inputs, ordered by date (older version of this data set with less inputs).
4) bank.csv with 10% of the examples and 17 inputs, randomly selected from 3 (older version of this data set with less inputs).